+++ parent-dep-mixin/index.js
@@ -1,5 +1,0 @@
-import parentDepMixin from './src/parent-dep-mixin';
-
-export {
-  parentDepMixin
-};